Fix del ruteo a login y registro
