+++ src/app/components/escaner/escaner.component.html
@@ -1,0 +1,5 @@
+<ion-fab class="ion-margin" vertical="top" horizontal="end" slot="fixed">
+  <ion-fab-button (click)="escanear()">
+    <ion-icon name="qr-code-sharp"></ion-icon>
+  </ion-fab-button>
+</ion-fab>
