+++ src/app/pages/tab1/tab1.page.html
@@ -1,7 +1,7 @@
 <ion-header [translucent]="true">
   <ion-toolbar color="dark">
     <ion-title>
-      Tab 1
+      Test
     </ion-title>
   </ion-toolbar>
 </ion-header>
@@ -71,13 +71,13 @@
     </ion-item>
 
     <ion-item>
-      <ion-button routerLink="/home/login" expand="block" fill="clear" shape="round">
+      <ion-button routerLink="/login" expand="block" fill="clear" shape="round">
         Iniciar Sesión
       </ion-button>
     </ion-item>
 
     <ion-item>
-      <ion-button routerLink="/home/register" expand="block" fill="clear" shape="round">
+      <ion-button routerLink="/register" expand="block" fill="clear" shape="round">
         Registro
       </ion-button>
     </ion-item>
